+++ src/main.js
@@ -38,10 +38,14 @@ import { store } from './store'
 
 /**
  *  Import languages
- * TODO: ¡¡¡¡¡¡¡UPDATE!!!!!!!!
  */
 // import i18n from './locales'
 import i18n from '@setup/i18n'
+
+/**
+ *  Import Geolocation plugin
+ */
+import VuexGeolocation from 'vuex-geolocation'
 
 /**
  * Import Timer
@@ -95,6 +99,12 @@ Vue.config.productionTip = process.env.NODE_ENV === 'production'
  * Load Onsen
  */
 Vue.use(VueOnsen)
+
+/**
+ * Load vuex-geolocation
+ */
+const vuexGeolocation = VuexGeolocation.sync(store)
+Vue.use(vuexGeolocation)
 
 /**
  * Load  Vue-Timer