+++ src/store/index.js
@@ -5,7 +5,6 @@ import alerts from '@modules/alerts'
 import auth from '@modules/auth'
 import errors from '@modules/errors'
 import localDataBase from '@modules/localDataBase'
-import location from '@modules/location'
 import navigator from '@modules/navigator'
 import shared from '@modules/shared'
 import social from '@modules/social'
@@ -20,7 +19,6 @@ export const store = new Vuex.Store({
     auth,
     errors,
     localDataBase,
-    location,
     navigator,
     shared,
     social,