+++ src/store/modules/location/getters.js
@@ -1,5 +1,0 @@
-import { USER_LOCATION } from '@store/types/getters_types'
-
-export default {
-  [USER_LOCATION]: state => state.userLocation
-}