+++ CHANGELOG.md
@@ -7,7 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.2.0] - 2019-05-15
+## [0.2.2] - 2019-05-16
+
+### Changed
+
+- Location page
+- Rename Location page -> UserLocation
+
+### Added
+
+- vuex-geolocation plugin
+- Location component
+
+### Removed
+
+- location store
+
+## [0.2.1] - 2019-05-15
 
 ### Fixed